requirements.txt
asgiref==3.9.1
Django==5.2.5
django-widget-tweaks==1.5.0
mysqlclient==2.2.7
sqlparse==0.5.3
tzdata==2025.2
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
djangorestframework==3.15.2